refactor(S-3.01): extract handle_remove to auth/remove.rs
Move handle_remove and handle_remove_in_memory out of mod.rs.
Re-export handle_remove as pub; re-export handle_remove_in_memory
under #[cfg(test)] so the tests in mod.rs can reference it via
super::*. 612 unit tests pass.
